extracted util functions to `utils/cssUtils.tsx`
