+++ src/components/common/Section.tsx
@@ -4,6 +4,8 @@ import { createStyles, Theme, TypographyTypeMap, withStyles, WithStyles } from "
 import Typography from "@material-ui/core/Typography";
 import Divider from "@material-ui/core/Divider";
 
+import { getStylesToFixPageHeaderOverlappingInPageAnchorIssue } from "../../utils/cssUtils";
+
 const styles = (theme: Theme) =>
     createStyles({
         section: {
@@ -52,29 +54,6 @@ function optionalDivider(hasDivider?: boolean) {
     return hasDivider
         ? <Divider/>
         : "";
-}
-
-function getStylesToFixPageHeaderOverlappingInPageAnchorIssue(theme: Theme) {
-    // To fix the AppBar overlapping in-page anchors issue
-    // Reference: https://stackoverflow.com/a/13117744
-    const padding = getPaddingTopForSection(theme);
-    return {
-        paddingTop: padding,
-        marginTop: -padding,
-    }
-}
-
-function getPaddingTopForSection(theme: Theme): string | number {
-    const minHeight = theme.mixins.toolbar.minHeight;
-    if (minHeight) {
-        if (typeof minHeight === "number") {
-            return theme.spacing(3) + minHeight;
-        } else {
-            return minHeight;
-        }
-    } else {
-        return theme.spacing(3);
-    }
 }
 
 Section.defaultProps = {
